Fix header parsing

+++ cdm-rest-api-client-generator-php/src/main/java/org/codingmatters/rest/php/api/client/model/ApiTypesPhpGenerator.java
@@ -27,7 +27,7 @@ public class ApiTypesPhpGenerator {
         typeMapping.put( "integer", "int" );
         typeMapping.put( "string", "string" );
         typeMapping.put( "boolean", "bool" );
-        typeMapping.put( "number", "int" );
+        typeMapping.put( "number", "float" );
         typeMapping.put( "datetime", "tz-datetime" );
         typeMapping.put( "datetime-only", "datetime" );
         typeMapping.put( "time-only", "time" );
